Rename android_hardware_smoke_test to androidhardwaresmoketest
