--- dev/integration_tests/android_hardware_smoke_test/.metadata
@@ -1,30 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
-
-version:
-  revision: "e209bda61adba4e7cb8b302cba01a656c70b603f"
-  channel: "[user-branch]"
-
-project_type: app
-
-# Tracks metadata for the flutter migrate command
-migration:
-  platforms:
-    - platform: root
-      create_revision: e209bda61adba4e7cb8b302cba01a656c70b603f
-      base_revision: e209bda61adba4e7cb8b302cba01a656c70b603f
-    - platform: android
-      create_revision: e209bda61adba4e7cb8b302cba01a656c70b603f
-      base_revision: e209bda61adba4e7cb8b302cba01a656c70b603f
-
-  # User provided section
-
-  # List of Local paths (relative to this file) that should be
-  # ignored by the migrate tool.
-  #
-  # Files that are not part of the templates will be ignored by default.
-  unmanaged_files:
-    - 'lib/main.dart'
-    - 'ios/Runner.xcodeproj/project.pbxproj'
