--- dev/integration_tests/android_hardware_smoke_test/android/app/src/main/AndroidManifest.xml
@@ -1,33 +1,0 @@
-<!-- Copyright 2014 The Flutter Authors. All rights reserved.
-Use of this source code is governed by a BSD-style license that can be
-found in the LICENSE file. -->
-
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <!-- ${applicationName} is used by the Flutter tool to select the Application
-         class to use. For most apps, this is the default Android application.
-         In most cases you can leave this as-is, but you if you want to provide
-         additional functionality it is fine to subclass or reimplement
-         Application and put your custom class here. -->
-    <application
-        android:name="${applicationName}"
-        android:label="android_hardware_smoke_test"
-        android:icon="@mipmap/ic_launcher">
-        <activity
-            android:name=".MainActivity"
-            android:theme="@style/LaunchTheme"
-            android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection|fontScale|screenLayout|density|uiMode"
-            android:hardwareAccelerated="true"
-            android:exported="true"
-            android:windowSoftInputMode="adjustResize">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <meta-data android:name="io.flutter.embedding.android.EnableImpeller" android:value="true" />
-        <meta-data android:name="io.flutter.embedding.android.ImpellerBackend" android:value="vulkan" />
-        <meta-data android:name="io.flutter.embedding.android.EnableHcpp" android:value="true" />
-        <meta-data android:name="flutterEmbedding" android:value="2" />
-    </application>
-</manifest>
