--- dev/integration_tests/android_hardware_smoke_test/integration_test/integration_test_wrapper.dart
@@ -1,55 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'dart:async';
-import 'dart:convert';
-
-import 'package:android_driver_extensions/extension.dart';
-import 'package:android_hardware_smoke_test/constants.dart';
-import 'package:android_hardware_smoke_test/main.dart' as app;
-import 'package:flutter/services.dart';
-import 'package:flutter_driver/driver_extension.dart';
-
-void main() {
-  enableFlutterDriverExtension(
-    // Register nativeDriverCommands so that host-side driver extensions can send NativeCommands
-    // (specifically for capturing system-level screenshots from the host during host-driven tests).
-    commands: <CommandExtension>[nativeDriverCommands],
-    // Thin handler to bridge driver's requestData and MainApp's test_channel.
-    handler: (String? request) async {
-      if (request == null) {
-        return json.encode(<String, Object?>{'message': 'Error: request was null'});
-      }
-
-      final Map<String, Object?> payload = (json.decode(request) as Map<Object?, Object?>)
-          .cast<String, Object?>();
-
-      // Handle host-side graphics backend self-discovery query
-      if (payload[keyCommand] == commandGetGoldenVariant) {
-        final String? variant = await const MethodChannel(
-          nativeSupportChannelName,
-        ).invokeMethod<String>(methodImpellerBackend);
-        return json.encode(<String, Object?>{keyGoldenVariant: variant});
-      }
-
-      // The request is encoded JSON, but there is no need to decode it here.
-      final ByteData message = const StringCodec().encodeMessage(request)!;
-      final completer = Completer<String>();
-
-      // ignore: deprecated_member_use
-      ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage(
-        testChannelName,
-        message,
-        (ByteData? replyData) {
-          final reply = const JSONMessageCodec().decodeMessage(replyData) as Map<Object?, Object?>?;
-          completer.complete(json.encode(reply));
-        },
-      );
-
-      return completer.future;
-    },
-  );
-
-  app.main();
-}
