+++ dev/integration_tests/androidhardwaresmoketest/android/app/build.gradle.kts
