+++ dev/integration_tests/androidhardwaresmoketest/android/app/src/main/kotlin/com/example/androidhardwaresmoketest/NativeTextView.kt
