+++ dev/integration_tests/androidhardwaresmoketest/android/app/test_result_embedder.gradle.kts
