+++ dev/integration_tests/androidhardwaresmoketest/android/gradle.properties
