+++ dev/integration_tests/androidhardwaresmoketest/android/settings.gradle.kts
