+++ dev/integration_tests/androidhardwaresmoketest/lib/backdrop_filter_blur.dart
@@ -1,0 +1,62 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'dart:ui' as ui;
+import 'package:flutter/material.dart';
+
+/// A custom widget displaying a colorful background canvas overlayed with a
+/// centered BackdropFilter Gaussian blur box.
+class BackdropFilterBlur extends StatelessWidget {
+  const BackdropFilterBlur({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      children: <Widget>[
+        const Positioned.fill(child: CustomPaint(painter: _BackgroundPainter())),
+        Center(
+          child: ClipRect(
+            child: BackdropFilter(
+              filter: ui.ImageFilter.blur(sigmaX: 5.0, sigmaY: 5.0),
+              child: Container(
+                width: 100,
+                height: 100,
+                alignment: Alignment.center,
+                decoration: BoxDecoration(
+                  color: Colors.white.withValues(alpha: 0.1),
+                  border: Border.all(color: Colors.white.withValues(alpha: 0.2)),
+                ),
+                child: const Text(
+                  'Blur',
+                  style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
+                ),
+              ),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+class _BackgroundPainter extends CustomPainter {
+  const _BackgroundPainter();
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    canvas.drawRect(Rect.fromLTWH(0, 0, size.width, size.height), Paint()..color = Colors.yellow);
+    canvas.drawCircle(
+      Offset(size.width * 0.3, size.height * 0.3),
+      60.0,
+      Paint()..color = Colors.red,
+    );
+    canvas.drawRect(
+      Rect.fromLTWH(size.width * 0.4, size.height * 0.4, 80, 80),
+      Paint()..color = Colors.blue,
+    );
+  }
+
+  @override
+  bool shouldRepaint(covariant CustomPainter oldDelegate) => false;
+}
