+++ dev/integration_tests/androidhardwaresmoketest/lib/image_drawing_canvas.dart
