+++ dev/integration_tests/androidhardwaresmoketest/lib/pixel_exact_local_file_comparator.dart
