+++ dev/integration_tests/androidhardwaresmoketest/lib/vector_drawings_canvas.dart
