+++ dev/integration_tests/androidhardwaresmoketest/pubspec.yaml
@@ -1,0 +1,47 @@
+name: androidhardwaresmoketest
+description: "A new Flutter project."
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+
+version: 1.0.0+1
+
+environment:
+  sdk: ^3.11.0-0
+resolution: workspace
+
+dependencies:
+  flutter:
+    sdk: flutter
+
+  path_provider: any
+  path: any
+  android_driver_extensions:
+    path: ../../tools/android_driver_extensions
+  flutter_driver:
+    sdk: flutter
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  test: any
+  image: any
+
+  # The "flutter_lints" package below contains a set of recommended lints to
+  # encourage good coding practices. The lint set provided by the package is
+  # activated in the `analysis_options.yaml` file located at the root of your
+  # package. See that file for information about deactivating specific lint
+  # rules and activating additional ones.
+  flutter_lints: any
+  integration_test:
+    sdk: flutter
+
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
+
+# The following section is specific to Flutter packages.
+flutter:
+  uses-material-design: true
+  assets:
+    - test_driver/goldens/
+
+
+# PUBSPEC CHECKSUM: tqljka
