+++ dev/integration_tests/androidhardwaresmoketest/test_driver/goldens/README.md
